Pin dagger CLI to v0.2
Since we published https://github.com/dagger/dagger/releases/tag/v0.3.6
(Nov. 30, 2022), this action started downloading that version (i.e.
`latest`), which resulted in breakage, as described in the issue that
this fixes.

As soon as all tests pass and this gets the OK + merge, we should:
1. Cut a new release for this action (I am thinking v3.2.0)
2. Update the v3 tag to point to this new version

That should fix this for all existing GitHub Actions users without any
intervention on their part.

Fixes:
- https://github.com/dagger/dagger-for-github/issues/80

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>