fix(ux): increase readability of install step
The official github self-hosted action container does not include curl in the OS
but because the `curl` command in the GH Action's install step redirects stderr
to /dev/null, it was hard to determine that.

As part of my Dagger interview, @gerhard and I investigated and found that issue.
We changed the `curl` flags from `-sL` to `-fsS` to be more verbose with the
error reporting and remove following redirects, as the move to Cloudfront
obviated the need for that flag.

We also removed the subshell in the install step at the end because it was not
needed and introduced extra complexity.

I could also see checking for the presence of curl at the beginning of the
script and failing quickly with a message, but that was not discussed in the
interview. Given that this happens very quickly and solves for other HTTP errors
it doesn't seem particularly necessary but I could see the potential for it, if
the desire for it was there.

Signed-off-by: Josh Ghiloni <ghiloni@gmail.com>